+++ library/QATools/QATools/PageObject/Proxy/WebElementCollectionProxy.php
@@ -12,6 +12,8 @@ namespace QATools\QATools\PageObject\Proxy;
 
 
 use QATools\QATools\PageObject\Element\WebElementCollection;
+use QATools\QATools\PageObject\ElementLocator\IElementLocator;
+use QATools\QATools\PageObject\IPageFactory;
 
 /**
  * Class for lazy-proxy creation to ensure, that WebElementCollection are
@@ -23,6 +25,21 @@ use QATools\QATools\PageObject\Element\WebElementCollection;
  */
 class WebElementCollectionProxy extends WebElementProxy
 {
+
+	/**
+	 * Initializes proxy for WebElementCollection.
+	 *
+	 * @param IElementLocator $locator      Element selector.
+	 * @param IPageFactory    $page_factory Page factory.
+	 */
+	public function __construct(IElementLocator $locator, IPageFactory $page_factory = null)
+	{
+		if ( !$this->className ) {
+			$this->className = '\\QATools\\QATools\\PageObject\\Element\\WebElementCollection';
+		}
+
+		parent::__construct($locator, $page_factory);
+	}
 
 	/**
 	 * Offset to set.